Move other PDF-related functionality to `pdf` module
